--- tests/__init__.py
@@ -1,4 +1,0 @@
-from proxy_external import *
-from proxy_internal import *
-from service_external import *
-from service_internal import *
