+++ tests/proxy_external/__init__.py
@@ -1,0 +1,3 @@
+"""Test the plugit proxy from an external point of view (Using the standalone proxy and a simple plugit service)"""
+
+from test_proxy_with_service import *
