+++ tests/proxy_external/test_proxy_with_service.py
@@ -1,5 +1,3 @@
-"""Test the plugit proxy from an external point of view (Using the standalone proxy and a simple plugit service)"""
-
 import unittest
 from nose.tools import *
