+++ tests/proxy_internal/__init__.py
@@ -1,0 +1,3 @@
+"""Test internal functions and features of the proxy"""
+
+
