+++ tests/service_external/__init__.py
@@ -1,0 +1,3 @@
+"""Test the simple service from an external point of view (Using HTTP requests)"""
+
+from test_service_from_http import *
