+++ tests/service_internal/__init__.py
@@ -1,0 +1,1 @@
+"""Test internal functions and features of a service"""
